requirements.txt
dlib-bin
imutils
numpy
opencv-python
pillow
pygame
scipy
streamlit
ultralytics